Update README.md
This GitHub repository contains the code and resources for a virtual internship project referred by Forage. The project focuses on analyzing call center data to uncover trends and patterns using Power BI. It includes features such as creating interactive dashboards, generating reports, and providing insights and recommendations based on the analysis. Feel free to explore the repository and reach out with any questions or feedback.